changed configuration directory stucture - config.py now hold all configurations' descriptions
